Fix export/import and remove axios
Update plugin to v2 paths

remove axios and update got to latest (12^).

+++ src/export_import.ts
@@ -4,20 +4,19 @@ import fs from "node:fs/promises";
 import tar from "tar";
 import { bdsRoot } from "./platformPathManeger";
 
+async function exists(filePath: string) {
+  return fs.access(filePath).then(() => true).catch(() => false);
+}
+
 export type payload = {
   httpVersion?: string,
-  request?: {
-    method: string,
-    path: string,
-  },
-  response?: {
-    code: number,
-    text: string
-  },
+  request?: {method: string, path: string},
+  response?: {code: number, text: string},
   header?: {[key: string]: string|boolean|number},
   second?: payload,
-  body?: any,
   raw?: string,
+  socket?: net.Socket
+  body?: any,
 };
 
 export class payloadError extends Error {
@@ -32,12 +31,16 @@ const payloadRequest = /^(GET|POST|CONNECT|HEAD|PUT|DELETE)\s+(.*)\s+HTTP\/([0-9
 const payloadResponse = /HTTP\/([0-9\.]+)\s+([0-9]+)\s+([\w\S\s]+)/;
 const parseHeard = /^([0-9A-Za-z\._-\s@]+):([\w\S\s]+)/;
 async function parsePayload(input: string|Buffer|net.Socket): Promise<payload> {
+  const payloadBody: payload = {header: {}};
   let data = "";
   if (typeof input === "string") data = input; else {
     if (Buffer.isBuffer(input)) data = input.toString("utf8");
-    else data = await new Promise<string>(done => input.once("data", dataInput => done(dataInput.toString("utf8"))))
+    else {
+      payloadBody.socket = input;
+      data = await new Promise<string>(done => input.once("data", dataInput => done(dataInput.toString("utf8"))));
+    }
   }
-  const payloadBody: payload = {raw: data, header: {}};
+  payloadBody.raw = data;
   if (/^{.*}$/.test(data.replace(/\r?\n/, ""))) {
     payloadBody.body = JSON.parse(data);
     return payloadBody;
@@ -90,7 +93,7 @@ function stringifyPayload(socket: net.Socket, response: payload) {
   let message = "";
   if (response.request) message += `${response.request.method.toUpperCase()} ${response.request.path} HTTP/${response.httpVersion||"1.0"}\r\n`;
   else message += `HTTP/${response.httpVersion||"1.0"} ${response.response.code} ${response.response.text}\r\n`;
-  if (response.header) Object.keys(response.header).forEach(key => `${key}: ${response.header[key]}\r\n`);
+  if (response.header) message += (Object.keys(response.header).map(key => `${key}: ${response.header[key]}`).join("\r\n"))+"\r\n";
   message += "\r\n";
   if (response.body !== undefined) {
     if (Array.isArray(response.body)||typeof response.body === "object") message += JSON.stringify(response.body);
@@ -108,7 +111,6 @@ export class exportBds {
   #server = net.createServer(async (socket): Promise<any> => {
     if (!this.acceptConnection) return stringifyPayload(socket, {response: {code: 400, text: "Server locked"}, body: {erro: "Server locked"}}).end();
     const payload = await parsePayload(await new Promise<string>(done => socket.once("data", res => done(res.toString("utf8")))));
-    console.log(payload);
     if (payload.header.Authorization !== this.authToken) return stringifyPayload(socket, {response: {code: 401, text: "Not allowed"}, body: {error: "Invalid token"}}).end();
     else stringifyPayload(socket, {response: {code: 200, text: "Success"}, header: {Date: (new Date()).toISOString(), "Content-Type": "bdsStream/tar"}});
     this.acceptConnection = false;
@@ -116,8 +118,8 @@ export class exportBds {
     // Compact bds root
     const tarCompress = tar.create({gzip: true, cwd: bdsRoot}, await fs.readdir(bdsRoot));
     tarCompress.pipe(socket);
-    tarCompress.on("data", ({length}) => console.log("Send to %s, size: %f", socket.localAddress+":"+socket.localPort, length));
     tarCompress.on("end", () => this.#server.close());
+    socket.on("close", tarCompress.destroy);
   });
 
   public async listen(port = 0) {
@@ -139,15 +141,14 @@ export class exportBds {
 }
 
 export async function importBds(option: {host: string, port: number, authToken: string}) {
-  await fs.rename(bdsRoot, bdsRoot+"_backup_"+Date.now());
-  const client = stringifyPayload(net.createConnection({host: option.host, port: option.port}), {request: {method: "GET", path: "/"}, header: {Authorization: option.authToken}});
-  await parsePayload(client);
+  if (await exists(bdsRoot)) await fs.rename(bdsRoot, bdsRoot+"_backup_"+Date.now());
+  const {socket} = await parsePayload(stringifyPayload(net.createConnection({host: option.host, port: option.port}), {request: {method: "GET", path: "/"}, header: {Authorization: option.authToken}}));
+  console.info("Connection sucess!");
   const tar_extract = tar.extract({cwd: bdsRoot, noChmod: false, noMtime: false, preserveOwner: true});
-  client.pipe(tar_extract);
-  client.on("data", ({length}) => console.log("Recive size: %f", length));
+  socket.pipe(tar_extract);
   return new Promise<void>((done, reject) => {
-    client.once("close", () => done());
-    client.on("error", reject);
+    socket.on("error", reject);
     tar_extract.on("error", reject);
+    socket.once("close", () => tar_extract.on("finish", done));
   });
 }
